requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
datasets==2.15.0
dill==0.3.7
distro==1.8.0
dnspython==2.4.2
evaluate==0.4.1
exceptiongroup==1.2.0
fastapi==0.104.1
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.19.4
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.0
numpy==1.26.2
openai==1.6.0
packaging==23.2
pandas==2.1.4
Pillow==9.3.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
responses==0.18.0
safetensors==0.4.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tokenizers==0.15.0
torch
torchaudio
torchvision
tqdm==4.66.1
transformers==4.36.2
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
xxhash==3.4.1
yarl==1.9.4
